chore(views): move all setting syncing logic to dedicated page hooks
