+++ shared/src/containers/Views/hooks/index.ts
@@ -1,0 +1,9 @@
+// Unified page-specific view settings hooks
+export { useOverviewViewSettings } from './useOverviewViewSettings'
+export { useTaskProgressViewSettings } from './useTaskProgressViewSettings'
+
+// Other hooks
+export { default as useBuildViewMenuItems } from './useBuildViewMenuItems'
+export { useSaveViewFromCurrent } from './useSaveViewFromCurrent'
+export { useSelectedView } from './useSelectedView'
+export { useViewsMutations } from './useViewsMutations'
